Squashed commit of the first purge
commit 1dea0f670b7c5c6e39a9724dfe330813c02ca8c9
Author: David <davidkazlauskas@gmail.com>
Date:   Mon Feb 2 17:54:59 2015 +0200

    update

    ...

commit aeaeee6ca9e981fc646f56d6e29a7dc7c632aae6
Author: David <davidkazlauskas@gmail.com>
Date:   Thu Jan 29 18:22:31 2015 +0200

    job

+++ CollectionAdapter.h
@@ -41,7 +41,7 @@ struct CollectionAdapter {
     static ValueType& getByIndex(ThisCol& c, size_t i);
     static ConstValueType& getByIndex(ConstCol& c, size_t i);
 
-    static int size(ConstCol& c);
+    static long size(ConstCol& c);
 
     static void erase(ThisCol& c, Iterator i);
     static void erase(ThisCol& c, Iterator beg, Iterator end);
