+++ proxy/Skipper.h
@@ -135,13 +135,13 @@ public:
             static const bool isNaiveAdvance =
                 !util::IsRandomAccessIteratorTagged<decltype(i)>::value;
             typedef detail::AdvancePicker<isNaiveAdvance> A;
-            int mul = _sk * ProxUtil::get_mul(_c.getRef());
+            long mul = _sk * ProxUtil::get_mul(_c.getRef());
             A::adv(i,e,mul * n);
             return Iterator(*this,i,_sk);
         }
     }
 
-    int size() const {
+    long size() const {
         if (!_cleared) {
             return -1;
         } else {
@@ -241,7 +241,7 @@ public:
         return ProxUtil::unwrap(_c.getRef());
     }
 
-    int getMul() const {
+    long getMul() const {
         return ProxUtil::get_mul(_c.cgetRef());
     }
 
@@ -288,7 +288,7 @@ struct IsProxy< Skipper< T,StoragePolicy > > {
         return c.getInternal();
     }
 
-    typedef int Dist;
+    typedef long Dist;
     template <class U>
     static Dist get_mul(U&& u) {
         return u.getMul();
@@ -379,7 +379,7 @@ struct CollectionAdapter< Skipper<T,StoragePolicy> > {
     }
 
     template <class C>
-    static int size(C&& c) {
+    static long size(C&& c) {
         return c.size();
     }
