+++ virtual/Socket.h
@@ -74,7 +74,7 @@ struct CollectionAdapter< Socket<T> > {
     }
 
     template <class U = int>
-    static int size(ConstCol& c) {
+    static long size(ConstCol& c) {
         // suppress static assert until method is actually called
         static_assert(templatious::util::DummyResolver<U, false>::val,
                       "Socket class can only be added to.");
