+++ virtual/modular/Tail.h
@@ -77,7 +77,7 @@ struct CollectionAdapter< ::templatious::vmodular::Tail<T> > {
         return c.cgetByIndex(i);
     }
 
-    static int size(ConstCol& c) {
+    static long size(ConstCol& c) {
         return c.size();
     }
 
@@ -178,7 +178,7 @@ struct CollectionAdapter< const ::templatious::vmodular::Tail<T> > {
         return c.cgetByIndex(i);
     }
 
-    static int size(ConstCol& c) {
+    static long size(ConstCol& c) {
         return c.size();
     }
